% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_network_path}
\alias{get_network_path}
\title{Get network path}
\usage{
get_network_path()
}
\value{
stored network path
}
\description{
This will retrieve the environment variable \code{SAFEPATHS_NETWORK_PATH}, which is stored in
your \code{.Renviron} file using \code{set_network_path()}.
}
\examples{
\dontrun{
get_network_path()
}
}
