% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safepaths_sitrep}
\alias{safepaths_sitrep}
\title{Check to see if a setpath is set and accessible}
\usage{
safepaths_sitrep()
}
\description{
This function can be helpful to detect instances where you might not have
the path set or are connected to VPN. If you are not connected or the path is
not set, the function will. Otherwise, the function will return \code{TRUE} invisibly.
}
