% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_network_path}
\alias{set_network_path}
\title{Set a network path}
\usage{
set_network_path(x)
}
\arguments{
\item{x}{character. A path to a folder on a network drive.
For Mac users it will look something like \code{"/Volumes/server_name/folder_name"}.
For Windows users, you can use the mapped drive letter or the UNC path
(e.g., \code{"P:/folder_name"} or \code{"//UNC_path"}).}
}
\value{
logical (invisibly)
}
\description{
This will add an environment variable called \code{SAFEPATHS_NETWORK_PATH} to
your \code{.Renviron} file and set the variable to the provided network
path.
}
\examples{
\dontrun{
# Mac
set_network_path("/Volumes/server_name/folder_name")
# Windows
set_network_path("P:/folder_name") OR
set_network_path("//UNC_path")
}
}
